+++ demo/src/main/java/raven/modal/demo/Demo.java
@@ -12,7 +12,7 @@ import java.awt.*;
 
 public class Demo extends JFrame {
 
-    public static final String DEMO_VERSION = "2.5.0";
+    public static final String DEMO_VERSION = "2.5.1";
 
     public Demo() {
         init();