+++ modal-dialog/src/main/java/raven/modal/component/ModalController.java
@@ -2,6 +2,8 @@ package raven.modal.component;
 
 import com.formdev.flatlaf.util.Animator;
 import com.formdev.flatlaf.util.CubicBezierEasing;
+import com.formdev.flatlaf.util.HiDPIUtils;
+import com.formdev.flatlaf.util.UIScale;
 import raven.modal.option.Option;
 import raven.modal.slider.PanelSlider;
 import raven.modal.utils.ImageSnapshots;
@@ -22,6 +24,7 @@ public class ModalController extends AbstractModalController {
     private boolean showing;
     private Animator animator;
     private float animated;
+    private double systemScaleFactor;
     private Image snapshotsImage;
 
     public ModalController(AbstractModalContainerLayer modalContainerLayer, ModalContainer modalContainer, Option option) {
@@ -32,7 +35,7 @@ public class ModalController extends AbstractModalController {
 
     protected boolean isUseAnimator(boolean show) {
         return option.isAnimationEnabled()
-                && (show == true || option.isAnimationOnClose()
+                && (show || option.isAnimationOnClose()
         );
     }
 
@@ -51,9 +54,10 @@ public class ModalController extends AbstractModalController {
                     @Override
                     public void begin() {
                         modalContainerLayer.animatedBegin();
+                        systemScaleFactor = UIScale.getSystemScaleFactor(getGraphicsConfiguration());
                         Border border = getBorder();
                         if (border != null) {
-                            snapshotsImage = ImageSnapshots.createSnapshotsImage(panelSlider, ModalController.this, getBorder());
+                            snapshotsImage = ImageSnapshots.createSnapshotsImage(panelSlider, ModalController.this, getBorder(), systemScaleFactor);
                         } else {
                             snapshotsImage = ImageSnapshots.createSnapshotsImage(panelSlider, 0);
                         }
@@ -178,31 +182,47 @@ public class ModalController extends AbstractModalController {
             if (snapshotsImage != null) {
                 Graphics2D g2 = (Graphics2D) g.create();
                 g2.setComposite(AlphaComposite.SrcOver.derive(animated));
-
-                float scaleValue = option.getLayoutOption().getAnimateScale();
-                if (scaleValue != 0) {
-                    float minScale = 1f - scaleValue;
-
-                    float scale = minScale + (scaleValue * animated);
-                    float width = getWidth();
-                    float height = getHeight();
-
-                    // calculate the center position after scaling
-                    float scaledWidth = (width * scale);
-                    float scaledHeight = (height * scale);
-                    float x = (width - scaledWidth) / 2f;
-                    float y = (height - scaledHeight) / 2f;
-                    g2.translate(x, y);
-                    g2.scale(scale, scale);
+                try {
+                    // draw snapshots image
+                    if (systemScaleFactor > 1) {
+                        HiDPIUtils.paintAtScale1x(g2, 0, 0, 100, 100, // width and height are not used
+                                (g2d, x2, y2, width2, height2, scaleFactor2) -> {
+                                    float scaleValue = option.getLayoutOption().getAnimateScale();
+                                    if (scaleValue != 0) {
+                                        scaleGraphics(g2d, scaleValue);
+                                    }
+                                    g2d.drawImage(snapshotsImage, x2, y2, null);
+                                });
+                    } else {
+                        float scaleValue = option.getLayoutOption().getAnimateScale();
+                        if (scaleValue != 0) {
+                            scaleGraphics(g2, scaleValue);
+                        }
+                        g2.drawImage(snapshotsImage, 0, 0, null);
+                    }
+                } finally {
+                    g2.dispose();
                 }
-
-                // draw snapshots image
-                g2.drawImage(snapshotsImage, 0, 0, null);
-                g2.dispose();
             } else {
                 super.paint(g);
             }
         }
+    }
+
+    private void scaleGraphics(Graphics2D g2, float scaleValue) {
+        float minScale = 1f - scaleValue;
+
+        float scale = minScale + (scaleValue * animated);
+        float width = getWidth();
+        float height = getHeight();
+
+        // calculate the center position after scaling
+        float scaledWidth = (width * scale);
+        float scaledHeight = (height * scale);
+        float x = (width - scaledWidth) / 2f;
+        float y = (height - scaledHeight) / 2f;
+        g2.translate(x, y);
+        g2.scale(scale, scale);
     }
 
     public float getAnimated() {