+++ modal-dialog/src/main/java/raven/modal/drawer/item/Item.java
@@ -103,7 +103,7 @@ public class Item implements MenuItem {
             return name;
         }
 
-        private String name;
+        private final String name;
 
         public Label(String name) {
             this.name = name;